Move RetryWithBackoff to utils package
Move the retry helper from the redis nonce manager package to pkg/utils/backoff.go
where it lives alongside the existing NewBackoff utility. This makes it reusable
across the codebase.
