+++ pkg/utils/backoff_test.go
@@ -1,8 +1,9 @@
-package redis
+package utils
 
 import (
 	"errors"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -10,10 +11,10 @@ import (
 
 func TestRetryWithBackoff_SucceedsFirstAttempt(t *testing.T) {
 	calls := 0
-	attempts, err := retryWithBackoff(func() error {
+	attempts, err := RetryWithBackoff(func() error {
 		calls++
 		return nil
-	})
+	}, 3, 50*time.Millisecond)
 	require.NoError(t, err)
 	assert.Equal(t, 1, attempts)
 	assert.Equal(t, 1, calls)
@@ -21,13 +22,13 @@ func TestRetryWithBackoff_SucceedsFirstAttempt(t *testing.T) {
 
 func TestRetryWithBackoff_SucceedsOnRetry(t *testing.T) {
 	calls := 0
-	attempts, err := retryWithBackoff(func() error {
+	attempts, err := RetryWithBackoff(func() error {
 		calls++
 		if calls < 2 {
 			return errors.New("transient error")
 		}
 		return nil
-	})
+	}, 3, 50*time.Millisecond)
 	require.NoError(t, err)
 	assert.Equal(t, 2, attempts)
 	assert.Equal(t, 2, calls)
@@ -35,26 +36,26 @@ func TestRetryWithBackoff_SucceedsOnRetry(t *testing.T) {
 
 func TestRetryWithBackoff_ExhaustsRetries(t *testing.T) {
 	calls := 0
-	attempts, err := retryWithBackoff(func() error {
+	attempts, err := RetryWithBackoff(func() error {
 		calls++
 		return errors.New("persistent error")
-	})
+	}, 3, 50*time.Millisecond)
 	require.Error(t, err)
-	assert.Equal(t, maxRetries, attempts)
-	assert.Equal(t, maxRetries, calls)
+	assert.Equal(t, 3, attempts)
+	assert.Equal(t, 3, calls)
 	assert.Contains(t, err.Error(), "persistent error")
 }
 
 func TestRetryWithBackoff_SucceedsOnLastAttempt(t *testing.T) {
 	calls := 0
-	attempts, err := retryWithBackoff(func() error {
+	attempts, err := RetryWithBackoff(func() error {
 		calls++
-		if calls < maxRetries {
+		if calls < 3 {
 			return errors.New("transient error")
 		}
 		return nil
-	})
+	}, 3, 50*time.Millisecond)
 	require.NoError(t, err)
-	assert.Equal(t, maxRetries, attempts)
-	assert.Equal(t, maxRetries, calls)
+	assert.Equal(t, 3, attempts)
+	assert.Equal(t, 3, calls)
 }
